Move composite-checkout tests to be runnable
At some point the jest configuration must have changed which prevented
running tests in subdirectories of `test/`. This moves all the tests up
one level so that they will be run.

It also removes the (no longer necessary) `babel/polyfill` require.
